took out of client based on feedback and rescoped it
